requirements.txt
crowdstrike_falconpy==1.2.12
tabulate==0.9.0
termcolor==2.2.0
